feat(api) user can update their password.
- sending email with mailtrap.io for more info: https://mailtrap.io/blog/spring-send-email/#How-to-send-email-in-Java-Spring-Boot-and-SMTP
I wasn't sure how to apporach this so i went with the following:

The user can visit a link and request a user reset email, the email will provide a link with a uuid. This uuid is valid for 60 minutes. if it is expired then the endpoint doesnt work anymore. through that endpoint the user can reset their password. They can have 5 active password reset requests, if this limit gets exceeded they have to wait for 60 minutes.

On top of that i adjusts some minor things:
- reset the terminal colors to normal after banner
- removed null checks form organization dto's
- adjusted user class so that equals checks on id instead of same instance.
- fixed flawed logic in validator class